requirements.txt
reflex==0.8.24.post1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
bcrypt==4.0.1
passlib==1.7.4
pydantic[email]==2.8.0